Create reducer generator with base logic
