+++ src/redux/specialAnime/specialAnime.js
@@ -1,0 +1,7 @@
+import generateReducer from '../util/reducerGenerator';
+
+const { fetchAnime: fetchSpecialAnime, reducer } = generateReducer('specialAnime', 'special');
+
+export { fetchSpecialAnime };
+
+export default reducer;
